DROTH-1196 (US 1174) Service: Override the Create and Update methods to support more than one property - Created New Dao file to MultiValueLinearAsset class.
